+++ Changelog.md
@@ -3,6 +3,7 @@
 ## master (unreleased)
 ### New features
 * New method `BugzillaHelper#bug_exists?`
+* Extract  `BugzillaHelper#bug_data` to separate method
 
 ## 0.1.1 (2018-05-15)
 * Fix getting id from string with spaces